fix: redesigned /data/{id} by merging info and preview
