+++ .changeset/friendly-glasses-agree.md
@@ -1,0 +1,5 @@
+---
+"geohub": patch
+---
+
+fix: redesigned /data/{id} by merging info and preview
